renames to MyPortfolioWidget
