+++ app/components/dashboard/Dashboard.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Col, Row } from "reactstrap";
 import { LayoutAuthorized } from "../layouts/LayoutAuthorized";
 import { MessageSignModal } from "../modals/SignMessageModal";
-import { MyPortfolio } from "./myPortfolio/MyPortoflioWidget";
+import { MyPortfolio } from "./myPortfolio/MyPortfolioWidget";
 import { NotificationWidget } from "./notification-widget/NotificationWidget";
 import { UserInfo } from "./UserInfo";
